+++ incremental/src/main/java/net/automatalib/incremental/dfa/IncrementalDFABuilder.java
@@ -17,10 +17,8 @@ package net.automatalib.incremental.dfa;
 
 import net.automatalib.alphabet.SupportsGrowingAlphabet;
 import net.automatalib.automaton.fsa.DFA;
-import net.automatalib.graph.Graph;
 import net.automatalib.incremental.ConflictException;
 import net.automatalib.incremental.IncrementalConstruction;
-import net.automatalib.ts.UniversalDTS;
 import net.automatalib.word.Word;
 
 /**
@@ -75,9 +73,6 @@ public interface IncrementalDFABuilder<I> extends IncrementalConstruction<DFA<?,
     }
 
     @Override
-    Graph<?, ?> asGraph();
-
-    @Override
-    UniversalDTS<?, I, ?, Acceptance, Void> asTransitionSystem();
+    DFA<?, I> asTransitionSystem();
 
 }