+++ incremental/src/test/java/net/automatalib/incremental/dfa/AbstractIncrementalPCDFABuilderTest.java
@@ -185,14 +185,14 @@ public abstract class AbstractIncrementalPCDFABuilderTest {
         testTSViewInternal(incPcDfa.asTransitionSystem());
     }
 
-    private static <S> void testTSViewInternal(UniversalDTS<S, Character, ?, Acceptance, Void> view) {
+    private static <S> void testTSViewInternal(UniversalDTS<S, Character, ?, Boolean, Void> view) {
         final S s1 = view.getState(W_1);
         final S s2 = view.getState(W_2);
         final S s3 = view.getState(W_3);
 
-        Assert.assertTrue(view.getStateProperty(s1).toBoolean());
-        Assert.assertFalse(view.getStateProperty(s2).toBoolean());
-        Assert.assertTrue(view.getStateProperty(s3).toBoolean());
+        Assert.assertTrue(view.getStateProperty(s1));
+        Assert.assertFalse(view.getStateProperty(s2));
+        Assert.assertTrue(view.getStateProperty(s3));
     }
 
     @Test(dependsOnMethods = "testLookup")